+++ ft_printf.h
@@ -54,9 +54,9 @@ void	ft_s_width_p(t_flags *prt, va_list args, size_t *count_char);
 
 void	ft_putnbru(unsigned int n, size_t *count_char);
 void	ft_putnbrdi(int n, size_t *count_char);
-void	ft_adres(unsigned long int nb, size_t *count_char);
-void	ft_putnbr_hexa(unsigned long int n, size_t *count_char);
-void	ft_putnbr_hexa_maj(unsigned long int n, size_t *count_char);
+void	ft_adres(unsigned int nb, size_t *count_char);
+void	ft_putnbr_hexa(unsigned int n, size_t *count_char);
+void	ft_putnbr_hexa_maj(unsigned int n, size_t *count_char);
 
 void	ft_full_flags(t_flags *ptr, size_t *count_char, va_list args);
 
@@ -67,7 +67,7 @@ int     ft_count_nbr(const char *str, int *i);
 int		ft_strlen(char *str);
 int		ft_unsigned_len(unsigned int nb);
 int		ft_intlen(int nb);
-int		ft_hexa_len(unsigned long int nb);
+int		ft_hexa_len(unsigned int nb);
 
 void	ft_s_left_wp(t_flags *prt, va_list args, size_t *count_char);
